+++ src/main/java/com/zipcodewilmington/scientificcalculator/Subtraction.java
@@ -1,0 +1,15 @@
+package com.zipcodewilmington.scientificcalculator;
+
+    public class Subtraction{
+    private double value1;
+    private double value2;
+
+    public Subtraction(){
+    }
+
+    public Double subtract(double value1, double value2){
+        return value1 - value2;
+    }
+}
+
+
