+++ src/test/java/com/zipcodewilmington/scientificcalculator/SubtractionTest.java
@@ -1,0 +1,17 @@
+package com.zipcodewilmington.scientificcalculator;
+
+import org.junit.Before;
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class SubtractionTest {
+
+    @Before
+    public void setUp() throws Exception {
+    }
+
+    @Test
+    public void subtract() {
+    }
+}
